requirements.txt
Flask==3.0.3
twilio
langchain==0.2.16
langchain-openai==0.1.22
python-dotenv==1.0.1
langchain-groq
requests
